core/pyproject.toml
[project]
name = "basecamp-core"
version = "0.7.3"
description = "Claude Code multi-project workspace launcher"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "questionary>=2.1.1",
    "rich-click>=1.8",
]

[project.scripts]
basecamp = "core.main:basecamp"

[dependency-groups]
dev = [
    "ruff~=0.13.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "pytest-env>=1.2.0",
    "factory-boy>=3.3.0",
    "freezegun>=1.5.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/core"]
